# CourseWebApplication/CourseWebApplication/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Reader API",
    "version": "1"
  },
  "paths": {
    "/api/Login": {
      "post": {
        "tags": [
          "Login"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Mail": {
      "post": {
        "tags": [
          "Mail"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SendMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SendMessageReponse"
            }
          }
        }
      }
    },
    "/api/Users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Users": {
      "put": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Users/registration/check": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CheckUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "CheckUserRequest": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "LoginRequest": {
      "required": [
        "login",
        "password"
      ],
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "RegisterRequest": {
      "required": [
        "email",
        "login",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "SendMessageReponse": {
      "required": [
        "code"
      ],
      "type": "object",
      "properties": {
        "code": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "SendMessageRequest": {
      "required": [
        "email"
      ],
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "UpdateRequest": {
      "required": [
        "password"
      ],
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}